ref(category): update api edit, remove, addCategory

+++ src/category/category.service.ts
@@ -32,6 +32,18 @@ export class CategoryService {
     return response;
   }
 
+  async validation(id: number, idProject: number) {
+    const checkCategory = await this.checkCategory(id);
+    if (!checkCategory) {
+      throw new NotFoundException();
+    }
+    const project = await this.projectService.getOneById(idProject);
+    if (!project) {
+      throw new NotFoundException('Project not Exist');
+    }
+    return true;
+  }
+
   async add(dto: AddCategoryDTO) {
     try {
       const category = this.categoryRepo.create(dto);
@@ -41,20 +53,16 @@ export class CategoryService {
     }
   }
   async edit(id: number, idProject: number, dto: EditCategoryDTO) {
-    const checkCategory = await this.checkCategory(id);
-    if (!checkCategory) {
-      throw new NotFoundException();
-    }
-    const checkProject = await this.projectService.checkProjectByID(idProject);
-    if (!checkProject) {
-      throw new NotFoundException('Project not found');
+    const validation = await this.validation(id, idProject);
+    if (!validation) {
+      return validation;
     }
     const categoryExistProject = await this.categoryRepo.countCategoryInProject(
       id,
       idProject,
     );
     if (!categoryExistProject) {
-      throw new BadRequestException('Category not exist Project');
+      throw new BadRequestException('Category not Exist In Project');
     }
     try {
       return await this.categoryRepo.update(id, dto);
@@ -63,7 +71,11 @@ export class CategoryService {
     }
   }
 
-  async remove(id: number) {
+  async remove(id: number, idProject: number) {
+    const validation = this.validation(id, idProject);
+    if (!validation) {
+      return validation;
+    }
     try {
       await this.categoryRepo.delete(id);
       return id;
@@ -81,20 +93,16 @@ export class CategoryService {
   }
 
   async addCategoryInProject(id: number, idProject: number) {
-    const checkCategory = await this.checkCategory(id);
-    if (!checkCategory) {
-      throw new NotFoundException();
-    }
-    const checkProject = await this.projectService.checkProjectByID(idProject);
-    if (!checkProject) {
-      throw new NotFoundException('Project not Exist');
+    const validation = this.validation(id, idProject);
+    if (!validation) {
+      return validation;
     }
     const categoryExistProject = await this.categoryRepo.countCategoryInProject(
       id,
       idProject,
     );
     if (categoryExistProject) {
-      throw new BadRequestException('Category exist Project');
+      throw new BadRequestException('Category Exist In Project');
     }
     try {
       return await this.categoryRepo.update(id, { projectID: idProject });
